Refactor build wheels script to try to get around LLVM repo flakyness.

+++ scripts/build_torch_wheels.sh
@@ -8,29 +8,97 @@ RELEASE_VERSION=$2  # rX.Y or nightly
 DEFAULT_PYTHON_VERSION=3.6
 DEBIAN_FRONTEND=noninteractive
 
+function setup_system {
+  sudo bash -c 'echo "APT::Acquire::Retries \"10\";" > /etc/apt/apt.conf.d/80-failparams'
+  sudo bash -c 'echo "APT::Acquire::http::Timeout \"180\";" >> /etc/apt/apt.conf.d/80-failparams'
+  sudo bash -c 'echo "APT::Acquire::ftp::Timeout \"180\";" >> /etc/apt/apt.conf.d/80-failparams'
+}
+
+function install_cudnn {
+  if [ "$CUDNN_TGZ_PATH" == "" ]; then
+    echo "Missing CUDNN_TGZ_PATH environment variable"
+    exit 1
+  fi
+  local CUDNN_FILE="cudnn.tar.gz"
+  if [[ "$CUDNN_TGZ_PATH" == gs://* ]]; then
+    gsutil cp "$CUDNN_TGZ_PATH" "$CUDNN_FILE"
+  else
+    wget -O "$CUDNN_FILE" "$CUDNN_TGZ_PATH"
+  fi
+  tar xvf "$CUDNN_FILE"
+  sudo cp cuda/include/cudnn.h /usr/local/cuda/include
+  sudo cp cuda/lib64/libcudnn* /usr/local/cuda/lib64
+  sudo chmod a+r /usr/local/cuda/include/cudnn.h /usr/local/cuda/lib64/libcudnn*
+  rm -f "$CUDNN_FILE"
+}
+
+function maybe_install_cuda {
+  if [ "$XLA_CUDA" == "1" ]; then
+    if [ ! -d "/usr/local/cuda" ]; then
+      local CUDA_VER="10.2"
+      local CUDA_SUBVER="89_440.33.01"
+      local CUDA_FILE="cuda_${CUDA_VER}.${CUDA_SUBVER}_linux.run"
+      wget "http://developer.download.nvidia.com/compute/cuda/${CUDA_VER}/Prod/local_installers/${CUDA_FILE}"
+      sudo sh "${CUDA_FILE}" --silent --toolkit
+      rm -f "${CUDA_FILE}"
+    fi
+    if [ ! -f "/usr/local/cuda/include/cudnn.h" ]; then
+      install_cudnn
+    fi
+  fi
+}
+
+function maybe_install_sources {
+  if [ ! -d "torch" ]; then
+    sudo apt-get install -y git
+    git clone --recursive https://github.com/pytorch/pytorch.git
+    cd pytorch
+    git clone --recursive https://github.com/pytorch/xla.git
+    export RELEASE_VERSION="nightly"
+  fi
+}
+
 function install_bazel() {
   local BAZEL_VERSION=$(cat "xla/third_party/tensorflow/.bazelversion")
   local BAZEL_FILE="bazel-${BAZEL_VERSION}-installer-linux-x86_64.sh"
   sudo apt-get install -y pkg-config zip zlib1g-dev unzip
+  # Move to /tmp folder as otherwise the .bazelversion file found in pytorch source makes
+  # the install fail.
+  pushd /tmp
   curl -L -O "https://github.com/bazelbuild/bazel/releases/download/${BAZEL_VERSION}/${BAZEL_FILE}"
   chmod 755 "$BAZEL_FILE"
   ./"$BAZEL_FILE" --user
+  rm -f "$BAZEL_FILE"
+  popd
   export PATH="$PATH:$HOME/bin"
 }
 
+function debian_version {
+  local VER
+  if ! sudo apt-get install -y lsb-release > /dev/null 2>&1 ; then
+    VER="buster"
+  else
+    VER=$(lsb_release -c -s)
+  fi
+  echo "$VER"
+}
+
 function install_llvm_clang() {
-  sudo apt-get -y install wget
-  sudo bash -c 'echo "deb http://apt.llvm.org/stretch/ llvm-toolchain-stretch-8 main" >> /etc/apt/sources.list'
-  sudo bash -c 'echo "deb-src http://apt.llvm.org/stretch/ llvm-toolchain-stretch-8 main" >> /etc/apt/sources.list'
-  wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key|sudo apt-key add -
-  sudo apt-get update
-  sudo apt-get -y install clang-8 clang++-8
+  local DEBVER=$(debian_version)
+  if ! apt-get install -y -s clang-8 > /dev/null 2>&1 ; then
+    sudo bash -c "echo \"deb http://apt.llvm.org/${DEBVER}/ llvm-toolchain-${DEBVER}-8 main\" >> /etc/apt/sources.list"
+    sudo bash -c "echo \"deb-src http://apt.llvm.org/${DEBVER}/ llvm-toolchain-${DEBVER}-8 main\" >> /etc/apt/sources.list"
+    wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key|sudo apt-key add -
+    sudo apt-get update
+  fi
+  sudo apt-get install -y clang-8 clang++-8
   echo 'export CC=clang-8 CXX=clang++-8' >> ~/.bashrc
   export CC=clang-8 CXX=clang++-8
 }
 
 function install_req_packages() {
-  sudo apt-get -y install python-pip git curl libopenblas-dev vim apt-transport-https ca-certificates
+  sudo apt-get -y install python-pip git curl libopenblas-dev vim apt-transport-https ca-certificates wget
+  maybe_install_cuda
   install_bazel
 }
 
@@ -45,9 +113,10 @@ function install_gcloud() {
 function install_and_setup_conda() {
   # Install conda if dne already.
   if ! test -d "$HOME/anaconda3"; then
-    curl -O https://repo.anaconda.com/archive/Anaconda3-5.2.0-Linux-x86_64.sh
-    sh Anaconda3-5.2.0-Linux-x86_64.sh -b
-    rm Anaconda3-5.2.0-Linux-x86_64.sh
+    CONDA_VERSION="5.3.1"
+    curl -O "https://repo.anaconda.com/archive/Anaconda3-${CONDA_VERSION}-Linux-x86_64.sh"
+    sh "Anaconda3-${CONDA_VERSION}-Linux-x86_64.sh" -b
+    rm -f "Anaconda3-${CONDA_VERSION}-Linux-x86_64.sh"
   fi
   echo ". $HOME/anaconda3/etc/profile.d/conda.sh" >> ~/.bashrc
   source "$HOME/anaconda3/etc/profile.d/conda.sh"
@@ -80,7 +149,7 @@ function build_and_install_torch() {
   # Only checkout dependencies once PT commit/branch checked out.
   git submodule update --init --recursive
   # Apply patches to PT which are required by the XLA support.
-  $(dirname $0)/apply_patches.sh
+  xla/scripts/apply_patches.sh
   python setup.py bdist_wheel
   pip install dist/*.whl
 }
@@ -108,6 +177,8 @@ function install_torchvision_from_source() {
 }
 
 function main() {
+  setup_system
+  maybe_install_sources
   install_req_packages
   install_llvm_clang
   install_and_setup_conda
